+++ bigmler/processing/args.py
@@ -50,7 +50,8 @@ RESOURCE_TYPES = ["source", "dataset", "model", "ensemble", "batch_prediction",
                   "anomaly_score", "batch_anomaly_score", "project", "sample",
                   "association", "logistic_regression", "deepnet", "script",
                   "library", "execution", "evaluation", "topic_model",
-                  "linear_regression", "fusion", "prediction", "pca"]
+                  "linear_regression", "fusion", "prediction", "pca",
+                  "external_connector"]
 
 STORED_MODELS = ["model_file", "ensemble_file", "logistic_file",
                  "cluster_file", "anomaly_file", "deepnet_file",
@@ -1017,6 +1018,15 @@ def get_output_args(api, command_args, resume):
     except AttributeError:
         pass
 
+    # Parses connection info for external connectors
+    try:
+        if command_args.connection_json:
+            command_args.connection_json_ = u.read_json(
+                command_args.connection_json)
+        else:
+            command_args.connection_json_ = {}
+    except AttributeError:
+        pass
 
     return {"api": api, "args": command_args}