+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+3.27.0 (2020-05-26)
+~~~~~~~~~~~~~~~~~~~
+
+- Adding bigmler connector subcommand.
+- Fixing issues with --resources-log and --clear-logs in Python 3.X.
+
 3.26.4 (2020-04-27)
 ~~~~~~~~~~~~~~~~~~~